# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "gshare API Support",
            "email": "gshare@austinbspencer.com"
        }
    },
    "paths": {
...
        "/v1/galleries": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all galleries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "get all galleries that exist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Gallery"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new gallery.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "create a new gallery",
                "parameters": [
                    {
                        "description": "New Gallery",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Gallery"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Gallery"
                        }
                    }
                }
            }
        },
        "/v1/galleries/id/{galleryID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get gallery by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "get a gallery by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gallery"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update gallery.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "update the gallery with given payload",
                "parameters": [
                    {
                        "description": "Updated Gallery",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.GalleryUpdate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gallery"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete gallery by given ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "remove gallery by given ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/galleries/id/{galleryID}/images": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update gallery images order.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "update gallery images order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Image IDs",
                        "name": "images",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload a new Image.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "upload a new Image to the gallery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "src",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Image"
                        }
                    }
                }
            }
        },
        "/v1/galleries/id/{galleryID}/images/zips": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create the zips of the gallery.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "create the zip files of the gallery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/v1/galleries/id/{galleryID}/qr-code": {
            "get": {
                "description": "Generate the QR Code for the gallery with the ID given.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gallery"
                ],
                "summary": "create a QR Code for the gallery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gallery ID",
                        "name": "galleryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gallery"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Gallery": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "event_date": {
                    "description": "The date the event occurred",
                    "type": "string"
                },
                "events": {
                    "description": "All events related to this gallery",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Event"
                    }
                },
                "expiration": {
                    "description": "Date the gallery expires and will no longer be visible",
                    "type": "string"
                },
                "featured_image": {
                    "description": "Featured image for the gallery",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Image"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "All images in the gallery",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Image"
                    }
                },
                "images_count": {
                    "description": "Total number of images in the gallery",
                    "type": "integer"
                },
                "live": {
                    "description": "Date the gallery goes live",
                    "type": "string"
                },
                "path": {
                    "description": "Path for the gallery on frontend",
                    "type": "string"
                },
                "protected": {
                    "description": "If the gallery is password protected",
                    "type": "boolean"
                },
                "public": {
                    "description": "Determines if the collection is visible on the homepage\nIf protected is true, then it won't show regardless",
                    "type": "boolean"
                },
                "title": {
                    "description": "Title of the gallery",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "zips_ready": {
                    "description": "ZipsReady will track if the zip files are created and up to date\nIf a new image is uploaded/deleted since zips were created will switch to false",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}